config/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.2.0"
description = "blog on Django"
authors = ["Alfir Tyakin <alfir7777@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.13"
Django = "4.2.7"
django-ckeditor = "6.7.0"
django-recaptcha = "3.0.0"
django-smtp-ssl = "1.0"
djangorestframework = "3.14.0"
gunicorn = "21.2.0"
Pillow = "10.1.0"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.0"
pytils = "0.4.1"
sentry-sdk = "0.10.2"

[tool.poetry.group.dev.dependencies]
flake8 = "6.1.0"
flake8-import-order = "0.18.2"
django-debug-toolbar = "4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
